CryptoCompanion/requirements.txt
discord~=2.3.2
dotenv~=0.9.9
python-dotenv~=1.1.1
PyNaCl~=1.5.0